LPF-1023 - switch from profile driven to property driven behaviour
LPF-1023 - switch from profile driven to property driven behaviour
