+++ src/main/resources/application-prod.yml
@@ -16,6 +16,9 @@ gpfd:
       username: ${MOJFIN_DEV_WRITE_USERNAME}
       password: ${MOJFIN_DEV_WRITE_PASSWORD}
       driver-class-name: oracle.jdbc.pool.OracleDataSource
+ 
+  s3:
+    use-template-store: true
 
 # Azure AD and Graph config
 spring:
