+++ src/main/resources/application-uat.yml
@@ -15,6 +15,9 @@ gpfd:
       password: ${MOJFIN_DEV_WRITE_PASSWORD}
       driver-class-name: oracle.jdbc.pool.OracleDataSource
 
+  s3:
+    use-template-store: true
+
 spring:
   cloud:
     azure:
